Improve test and rename files
